pyproject.toml
[project]
name = "campus-connect-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.2",
    "gradio>=6.9.0",
    "jq>=1.11.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
]
